Create Doc container in spaCy
Doc container in spaCy
The first step of a spaCy text processing pipeline is to convert a given text string into a Doc container, which stores the processed text. In this exercise, you'll practice loading a spaCy model, creating an nlp() object, creating a Doc container and processing a text string that is available for you.

en_core_web_sm model is already downloaded.

Instructions

Load en_core_web_sm and create an nlp object.
Create a doc container of the text string.
Create a list containing the text of each tokens in the doc container.