Create Doc similarity with spaCy
Doc similarity with spaCy
Semantic similarity is the process of analyzing multiple sentences to identify similarities between them. In this exercise, you will practice calculating semantic similarities of documents to a given document. The goal is to categorize a list of given reviews that are relevant to canned dog food.

The canned dog food category is stored at category. A sample of five food reviews has been provided for you in a list called texts. en_core_web_md is loaded as nlp.

Instructions

Create a documents list containing Doc containers of all texts.
Create a Doc container of the category and store it as category_document.
Iterate through documents and print the similarity scores of each Doc container and the category_document, rounded to three digits.