Create Analyzing pipelines in spaCy
Analyzing pipelines in spaCy
spaCy allows you to analyze a spaCy pipeline to check whether any required attributes are not set. In this exercise, you'll practice analyzing a spaCy pipeline. Earlier in the video, an existing en_core_web_sm pipeline was analyzed and the result was No problems found., in this instance, you will analyze a blank spaCy English model with few added components and observe results of the analysis.

The spaCy package is already imported for you to use.

Instructions 1/2

Load a blank spaCy English model as nlp.
Add tagger and entity_linker pipeline components to the blank model.
Analyze the nlp pipeline.